# specs/001-product-overview/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AQ Trading
  version: 0.1.0
paths:
...
  /api/agents/invoke:
    post:
      tags:
      - agents
      summary: Invoke Agent
      description: "Invoke an agent task.\n\nThis endpoint dispatches a task to the\
        \ specified agent role.\nThe agent runs as a subprocess and returns results\
        \ asynchronously.\n\nArgs:\n    request: The agent invocation request containing\
        \ role, task, and context\n\nReturns:\n    AgentResultResponse with the execution\
        \ result\n\nExample:\n    POST /api/agents/invoke\n    {\n        \"role\"\
        : \"researcher\",\n        \"task\": \"analyze market trends for AAPL\",\n\
        \        \"context\": {\"symbol\": \"AAPL\", \"timeframe\": \"1D\"}\n    }"
      operationId: invoke_agent_api_agents_invoke_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentInvokeRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResultResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/agents/results:
    get:
      tags:
      - agents
      summary: List Agent Results
      description: "List agent results with pagination.\n\nArgs:\n    limit: Maximum\
        \ number of results to return (default 50, max 1000)\n    offset: Number of\
        \ results to skip for pagination\n\nReturns:\n    AgentResultsListResponse\
        \ containing results and total count\n\nExample:\n    GET /api/agents/results?limit=10&offset=0\n\
        \    Returns first 10 agent results"
      operationId: list_agent_results_api_agents_results_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 0
          description: Maximum results to return
          default: 50
          title: Limit
        description: Maximum results to return
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Number of results to skip
          default: 0
          title: Offset
        description: Number of results to skip
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResultsListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/agents/results/{result_id}:
    get:
      tags:
      - agents
      summary: Get Agent Result
      description: "Get a specific agent result by ID.\n\nArgs:\n    result_id: The\
        \ UUID of the agent result\n\nReturns:\n    AgentResultResponse with the result\
        \ details\n\nRaises:\n    404: If the result is not found\n\nExample:\n  \
        \  GET /api/agents/results/550e8400-e29b-41d4-a716-446655440000"
      operationId: get_agent_result_api_agents_results__result_id__get
      parameters:
      - name: result_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          description: The agent result UUID
          title: Result Id
        description: The agent result UUID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResultResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AgentInvokeRequest:
      properties:
        role:
          $ref: '#/components/schemas/AgentRole'
          description: The agent role to invoke
        task:
          type: string
          maxLength: 500
          minLength: 1
          title: Task
          description: Task description
        context:
          additionalProperties: true
          type: object
          title: Context
          description: Context data for the agent
      type: object
      required:
      - role
      - task
      - context
      title: AgentInvokeRequest
      description: "Request schema for invoking an agent task.\n\nAttributes:\n  \
        \  role: The agent role to invoke (researcher, analyst, risk_controller, ops)\n\
        \    task: Description of the task for the agent to perform\n    context:\
        \ Additional context data for the agent"
    AgentResultResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        role:
          $ref: '#/components/schemas/AgentRole'
        task:
          type: string
          title: Task
        context:
          additionalProperties: true
          type: object
          title: Context
        result:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Result
        success:
          type: boolean
          title: Success
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        started_at:
          type: string
          format: date-time
          title: Started At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        duration_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Ms
      type: object
      required:
      - id
      - role
      - task
      - context
      - result
      - success
      - error
      - started_at
      - completed_at
      - duration_ms
      title: AgentResultResponse
      description: "Response schema for an agent result.\n\nAttributes:\n    id: Unique\
        \ identifier for the result\n    role: The agent role that executed the task\n\
        \    task: The task that was executed\n    context: Context data provided\
        \ to the agent\n    result: The result data (if successful)\n    success:\
        \ Whether the task completed successfully\n    error: Error message (if failed)\n\
        \    started_at: When the agent started execution\n    completed_at: When\
        \ the agent completed execution\n    duration_ms: Execution duration in milliseconds"
    AgentResultsListResponse:
      properties:
        results:
          items:
            $ref: '#/components/schemas/AgentResultResponse'
          type: array
          title: Results
        total:
          type: integer
          title: Total
      type: object
      required:
      - results
      - total
      title: AgentResultsListResponse
      description: "Response schema for listing agent results.\n\nAttributes:\n  \
        \  results: List of agent results\n    total: Total number of results (for\
        \ pagination)"
    AgentRole:
      type: string
      enum:
      - researcher
      - analyst
      - risk_controller
      - ops
      title: AgentRole
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError